+++ src/Uno.Extensions.Navigation.UI/IRouteRegistry.cs
@@ -1,0 +1,6 @@
+﻿namespace Uno.Extensions.Navigation;
+
+public interface IRouteRegistry
+{
+	IRouteRegistry Register(RouteMap route);
+}
