+++ src/Uno.Extensions.Navigation.UI/NavigatorFactoryBuilder.cs
@@ -1,0 +1,13 @@
+﻿using System;
+#if !WINUI
+#else
+using Microsoft.UI.Xaml;
+using Microsoft.UI.Xaml.Controls;
+#endif
+
+namespace Uno.Extensions.Navigation;
+
+public class NavigatorFactoryBuilder
+{
+    public Action<INavigatorFactory>? Configure { get; set; }
+}
