--- src/Uno.Extensions.Navigation.UI/Navigators/ICompositeNavigator.cs
@@ -1,5 +1,0 @@
-﻿namespace Uno.Extensions.Navigation.Navigators;
-
-public interface ICompositeNavigator
-{
-}
